HTML & list.js cleanup, formatting, code updates.
## This consists primarily of updates & formatting cleanup.

### `jQuery'

- Updated __jQuery__ to a newer version (_v3.1.1_ - while larger, it
does cache so it should only reload once a year)

## `list.js`
- __list.js__ has been cleaned up for consistency & formatting.
- The function 'buildNavigation' at Line #56 seemed to missing several
semicolons(';'). I did not see a reason for this, and adding them
appears to have caused no problems in code function. If there is some
reason for this, then please let me know to better my understanding.

## `index.html`
- Reformatted the html to better follow the latest HTML5 statandards

## `clang-format`
I use __clang-format__ for most of my code based on the '_Google API_'.
The following are the settings used:
``
---
Language: JavaScript
BasedOnStyle: Google
# Use 100 columns for JS.
ColumnLimit: 100
``