Encode directory paths passed in the `prefix` URL parameter
As list.js passes through the `prefix` parameter to the S3 request's
`prefix` parameter and S3 documentation calls out characters in a key
name that "likely need to be URL encoded", encode directory paths the
same as file paths for use in `href`, including not encoding
forwardslashes.

Refactor encoding and decoding of forwardslashes into a function now
that it's used more than once.